Finalização da Classe Produto + validações + métodos/regras de negócio

+++ controlePedidos/src/main/java/com/br/chocolatePontoVirgula/model/services/ProdutoService.java
@@ -2,6 +2,7 @@ package com.br.chocolatePontoVirgula.model.services;
 
 import com.br.chocolatePontoVirgula.model.entity.Produto;
 import com.br.chocolatePontoVirgula.model.form.ProdutoForm;
+import com.br.chocolatePontoVirgula.model.repository.ItemPedidoRepository;
 import com.br.chocolatePontoVirgula.model.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -11,6 +12,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,6 +22,9 @@ public class ProdutoService {
 
     @Autowired
     private ProdutoRepository produtoRepository;
+
+    @Autowired
+    private ItemPedidoRepository itemPedidoRepository;
 
     public void save(@Validated ProdutoForm produto) {
         Produto p = new Produto();
@@ -46,8 +52,17 @@ public class ProdutoService {
         }
     }
 
-    public void delete(Long id){
-        produtoRepository.deleteById(id);
+    //verificar antes de excluir um produto para ver se ele está na tabela itemPedido
+    public ResponseEntity<String> delete(Long id) throws URISyntaxException {
+        URI uri = new URI("http://localhost:8080/itens-pedidos/produtosnoitem/" + id);
+        ResponseEntity<List<Long>> qnt = ResponseEntity.created(uri).body(itemPedidoRepository.produtoHasPedido(id));
+
+        if(qnt.getBody() != null){
+            return ResponseEntity.badRequest().body("Existem pedidos com esse produto");
+        } else {
+            produtoRepository.deleteById(id);
+            return ResponseEntity.ok().body("Produto excluído");
+        }
     }
 
     public Page<Produto> findAll(Pageable pageable) {
@@ -82,4 +97,5 @@ public class ProdutoService {
         Produto produtoEstoque = produtoRepository.getById(produto.getId());
         return produtoEstoque.getQuantidadeEstoque();
     }
+
 }
